Add loading rendering

+++ src/time.c
@@ -6,7 +6,7 @@
 /*   By: ohakola+veilo <ohakola+veilo@student.hi    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/31 14:34:58 by veilo             #+#    #+#             */
-/*   Updated: 2020/11/16 13:56:34 by ohakola+vei      ###   ########.fr       */
+/*   Updated: 2020/11/20 00:10:47 by ohakola+vei      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,5 +27,5 @@ float			sin_time(float min, float max, float speed)
 	amplitude = (max - min) / 2.0;
 	zero_offset = min + amplitude;
 	return (amplitude *
-		sin((SDL_GetTicks() / 1000.0) * speed ) + zero_offset);
+		sin(((float)SDL_GetTicks() / 1000.0) * speed ) + zero_offset);
 }
